+++ common/src/main/java/me/cg360/mod/bridging/BridgingMod.java
@@ -14,13 +14,12 @@ public class BridgingMod {
 
     public static final String MOD_ID = "bridgingmod";
 
-    public static final ResourceLocation PLACEMENT_ICONS_TEXTURE = ResourceLocation.tryBuild(MOD_ID, "textures/gui/placement_icons.png");
-
     private static boolean configSuccessfullyInitialized = true;
 
     public static void init() {
         BridgingConfig.HANDLER.load();
     }
+
 
     public static boolean isConfigSuccessfullyInitialized() {
         return configSuccessfullyInitialized;