+++ common/src/main/java/me/cg360/mod/bridging/raytrace/PathTraversalHandler.java
@@ -115,7 +115,7 @@ public class PathTraversalHandler {
         LinkedList<Direction> validSides = new LinkedList<>();
 
         for(Direction direction: Direction.values()) {
-            Vector3f directionNormal = Vec3.atLowerCornerOf(direction.getNormal()).toVector3f();
+            Vector3f directionNormal = Vec3.atLowerCornerOf(direction.getUnitVec3i()).toVector3f();
 
             double similarity = viewDirection.dot(directionNormal);
 
@@ -158,7 +158,7 @@ public class PathTraversalHandler {
         }
 
 
-        BlockPos blockPlacingOffOf = placementTarget.offset(checkSide.getNormal());
+        BlockPos blockPlacingOffOf = placementTarget.offset(checkSide.getUnitVec3i());
 
         // Can't place off of air or liquids.
         if(level.isEmptyBlock(blockPlacingOffOf)) return false;