+++ changelog.md
@@ -1,3 +1,12 @@
+__**Changelog (v2.5.1 for 1.21.3)**__
+
+- Chopped up `bridgingmod:gui/placement_icons` into:
+  - `bridgingmod:gui/indicator/up`
+  - `bridgingmod:gui/indicator/horizontal`
+  - `bridgingmod:gui/indicator/down`
+- Fixed incorrect texture path for bridge-assist config image.
+- 1.21.3 support!
+
 __**Changelog (v2.5.0)**__
 
 - Revamped the config entirely w/ YetAnotherConfigLibrary